lora repeater mode
This adds a settings.toml option for the base station to repeat
packets (with hop counter to prevent packet storms!). I'm using
this so that I can put the LCD base station in a spot that's good
for watching the display but bad for hearing the sensor packets.
When I add a repeater in a spot that has better line of sight to
the sensor, it all works out. Sorta. My RSSI is still just on the
edge of usable. I should probably use a higher spreading factor
and maybe a longer preamble.