add ESP-NOW gateway and remote display options
This makes it possible to have a LoRa node forward packets as an
ESP-NOW broadcast and to have ESP-NOW remote displays that don't
need a LoRa module. This works either in combination with, or as
in addition to, the LoRa repeater option.

The settings.toml options are getting a bit complicated because
I'm exploring different hardware configurations using the parts I
have on hand. Based on the difficulty I'm having with getting
reliable LoRa reception with a ground floor base station display,
it seems like an elevated receiver or gateway will be necessary.

Hardware builds now supported by the current option set:
1. Sensor: Feather ESP32-S3 + RFM95W LoRa Featherwing + MCP9808 I2C
   temperature sensor
2. Base station with optional gateway or repeater sub-modes:
   Feather ESP32-S3 + RFM95W LoRa Featherwing + (optional) I2C LCD
3. ESP-NOW remote display (no LoRa):
   Metro ESP32-S3 + 2.8" TFT touch display (ILI9341)