adapt for QT Py ESP32-S3
These changes get the base station firmware running on a plain
QT Py ESP32-S3 in `ESPNOW_DISPLAY = 1` mode with no SPI or I2C
devices attached.

This is mostly a workaround for me not having an extra LoRa radio
module to use for firmware dev. With the weather between seasons,
it's very useful to monitor the greenhouse sensors, even if it's
just scrolling log lines on a terminal.

But, I do need to work on an improved display, and I want to do
that by way of an IRC bridge. I'm currently running my main LoRa
base station in `ESPNOW_GATEWAY = 1` mode. That lets me recieve
copies of the LoRa sensor data packets on boards with ESP-NOW
without needing a LoRa module.

The point of this is to give me an additional USB serial feed of
sensor data so I can leave my main LoRa base station set up for
logging to a file with tio. I'll use the QT Py for working on an
IRC bridge.